+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- feat(Saisie de scrutin): passage du custom widget en vue3 [#82](https://github.com/betagouv/Resultats-Elections-FPT/pull/82)
 - core: améliore la configuration des tests automatisés [#81](https://github.com/betagouv/Resultats-Elections-FPT/pull/81)
 - core: Met en place des tests e2e [#80](https://github.com/betagouv/Resultats-Elections-FPT/pull/80)
 - fix(Rattachement): corrige le rattachement des collectivités triées par ordre alphabétique [#79](https://github.com/betagouv/Resultats-Elections-FPT/pull/79)